Create endpoint for create room
