+++ SchoolManagement.Api/Resources/en-US.json
@@ -10,5 +10,6 @@
   "Parent relation is required" :  "Parent relation is required",
   "Parent job is required" : "Parent job is required" ,
   "Parent phone is required" : "Parent phone is required",
-  "Student created successfully": "Student created successfully"
+  "Student created successfully": "Student created successfully",
+  "Expiration time must be greater than or equal 5 minutes.": "Expiration time must be greater than or equal 5 minutes."
 }
