+++ SchoolManagement.Application/Features/Rooms/Commands/CreateRoom/CreateRoomCommand.cs
@@ -1,3 +1,4 @@
+using System.ComponentModel.DataAnnotations;
 using MediatR;
 using SchoolManagement.Application.Features.Rooms.Service;
 
@@ -7,6 +8,8 @@ public class CreateRoomCommand : IRequest<CreateRoomResponse>
 {
     public string ChannelName { get; set; } = string.Empty;
     public string UserId { get; set; } = string.Empty;
+    
+    [Range(300, int.MaxValue, ErrorMessage = "Expiration time must be greater than or equal 5 minutes.")] // from 5 min 
     public int ExpirationTimeInSeconds { get; set; } = 3600; // Default to 1 hour
 }
 public class CreateRoomResponse
@@ -14,29 +17,6 @@ public class CreateRoomResponse
     public string ChannelName { get; set; } = string.Empty;
     public string UserId { get; set; } = string.Empty;
     public string Token { get; set; } = string.Empty;
+    public DateTime expiresAt { get; set; }
 }
 
-public class CreateRoomCommandHandler : IRequestHandler<CreateRoomCommand, CreateRoomResponse>
-{
-    private readonly AgoraTokenService _agoraTokenService;
-
-    public CreateRoomCommandHandler(AgoraTokenService agoraTokenService)
-    {
-        _agoraTokenService = agoraTokenService;
-    }
-
-    public Task<CreateRoomResponse> Handle(CreateRoomCommand request, CancellationToken cancellationToken)
-    {
-        var token = _agoraTokenService.GenerateToken(
-            request.ChannelName,
-            request.UserId,
-            request.ExpirationTimeInSeconds);
-
-        return Task.FromResult(new CreateRoomResponse
-        {
-            ChannelName = request.ChannelName,
-            UserId = request.UserId,
-            Token = token
-        });
-    }
-}
