+++ SchoolManagement.Application/Features/Rooms/Commands/CreateRoom/CreateRoomCommandHandler.cs
@@ -1,0 +1,30 @@
+using MediatR;
+using SchoolManagement.Application.Features.Rooms.Service;
+
+namespace SchoolManagement.Application.Features.Rooms.Commands.CreateRoom;
+
+public class CreateRoomCommandHandler : IRequestHandler<CreateRoomCommand, CreateRoomResponse>
+{
+    private readonly AgoraTokenService _agoraTokenService;
+
+    public CreateRoomCommandHandler(AgoraTokenService agoraTokenService)
+    {
+        _agoraTokenService = agoraTokenService;
+    }
+
+    public Task<CreateRoomResponse> Handle(CreateRoomCommand request, CancellationToken cancellationToken)
+    {
+        var token = _agoraTokenService.GenerateToken(
+            request.ChannelName,
+            request.UserId,
+            request.ExpirationTimeInSeconds);
+
+        return Task.FromResult(new CreateRoomResponse
+        {
+            ChannelName = request.ChannelName,
+            UserId = request.UserId,
+            Token = token,
+            expiresAt = DateTime.UtcNow.AddSeconds(request.ExpirationTimeInSeconds)
+        });
+    }
+}
