+++ SchoolManagement.Application/SchoolManagement.Application.csproj
@@ -13,6 +13,10 @@
   </ItemGroup>
 
   <ItemGroup>
+    <PackageReference Include="Agora-Token-Generator" Version="1.0.1" />
+    <PackageReference Include="AgoraNET" Version="2.0.0" />
+    <PackageReference Include="agorartc" Version="3.4.6" />
+    <PackageReference Include="agora_rtc_sdk" Version="4.2.6.19-alpha.1" />
     <PackageReference Include="AutoMapper" Version="13.0.1" />
     <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="12.0.1" />
     <PackageReference Include="MediatR" Version="12.4.1" />
